# Conflicts:
#	Sugukuru/WebContent/sales/create_order.jsp
#	Sugukuru/WebContent/stock/order_list.jsp
#	SugukuruDB/src/er/sgkl_db_1.0.erm

+++ Sugukuru/src/dtd/OrderDetail.java
@@ -13,42 +13,42 @@ public class OrderDetail {
 	 * @param orderId int 受注ID
 	 */
 	public int orderId;
-	
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
 	 * @param num int 行番号
 	 */
 	public int num;
-	
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
 	 * @param productName String 商品ID
 	 */
 	public String productId;
-	
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
 	 * @param productName String 商品名
 	 */
 	public String productName;
-	
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
 	 * @param price int 商品単価
 	 */
 	public int price;
-	
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
 	 * @param cunsumntionTax int 消費税
 	 */
-	public int cunsumntionTax;
-	
+	public /*int*/double cunsumntionTax;
+
 	/**
 	 * @auther Tester
 	 * 2016/11/02
